[FIX] Gave testotestorone to Auth Tests, now checking values returned and chekcing much more cases like :
Attempting to sign up with missing fields.
Attempting to sign up with an invalid email address.
Attempting to sign up with a weak password.
Attempting to sign up with an email address that already exists in the database.
Attempting to log in with missing email or password.
Attempting to log in with an email address that does not exist in the database.
Attempting to log in with an incorrect password.